requirements.txt
# Framework Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Configuration
python-dotenv==1.0.0

# Hugging Face & ML
transformers==4.37.0
torch==2.1.2
pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.3

# Détection Fake News (modèles légers)
sentence-transformers==2.3.1

# Audio processing
librosa==0.10.1
soundfile==0.12.1

# Utilitaires
python-multipart==0.0.6
aiofiles==23.2.1